framework builds, fp and training not implemented

+++ xorgens.h
@@ -1,6 +1,6 @@
 /* xorgens.h for xorgens version 3.04, R. P. Brent 20060628 */
 
-typedef unsigned long UINT; /* Type for random 32 or 64-bit integer, 
+typedef unsigned UINT; /* Type for random 32 or 64-bit integer,
                                e.g. unsigned long, unsigned long long,
                                uint64_t, unsigned int or uint32_t */
